refactor: 필요없는 로직 제거
